Started removing MathDecidr
