+++ package.json
@@ -29,7 +29,6 @@
     "itemsholdr": "0.6.X",
     "mapscreatr": "0.6.X",
     "mapscreenr": "0.6.X",
-    "mathdecidr": "0.6.X",
     "menugraphr": "0.6.X",
     "modattachr": "0.6.X",
     "numbermakr": "0.6.X",
