+++ shenanigans.json
@@ -14,7 +14,6 @@
     "ItemsHoldr": "0.6.X",
     "MapsCreatr": "0.6.X",
     "MapScreenr": "0.6.X",
-    "MathDecidr": "0.6.X",
     "MenuGraphr": "0.6.X",
     "ModAttachr": "0.6.X",
     "NumberMakr": "0.6.X",
