+++ src/IFullScreenPokemon.ts
@@ -97,11 +97,6 @@ export interface IMathModuleSettings extends igamestartr.IMathModuleSettings {
      * Constants the IMathDecidr may use in equations.
      */
     constants: IStaticMathConstants;
-
-    /**
-     * Calculation Functions, keyed by name.
-     */
-    equations: IMathEquations;
 }
 
 /**
@@ -229,224 +224,6 @@ export interface IMathConstants extends IStaticMathConstants {
 }
 
 /**
- * Calculation Functions for an IMathDecidr, keyed by name.
- */
-export interface IMathEquations extends imathdecidr.IEquations {
-    /**
-     * Generates a new Pokemon with the given traits.
-     * 
-     * @param title   The type of Pokemon to create.
-     * @param level   The level of the new Pokemon (by default, 1).
-     * @param moves   What moves the Pokemon has (by default, generated from its type
-     *                and level).
-     * @param iv   What IV points the Pokemon should start with (by default, generated 
-     *             from the newPokemonIVs equation.
-     * @param ev   What EV points the Pokemon should start with (by default, generated
-     *             from the newPokemonEVs equation).
-     * @returns A newly created Pokemon.
-     */
-    newPokemon: (
-        this: imathdecidr.IMathDecidr,
-        title: string[],
-        level?: number,
-        moves?: ibattlemovr.IMove[],
-        iv?: number,
-        ev?: number) => IPokemon;
-
-    /**
-     * Computes the default new moves for a Pokemon based on its type and level.
-     * 
-     * @param title   The type of Pokemon.
-     * @param level   The level of the Pokemon.
-     * @returns The default moves of the Pokemon.
-     */
-    newPokemonMoves: (this: imathdecidr.IMathDecidr, title: string[], level: number) => ibattlemovr.IMove[];
-
-    /**
-     * Computes a random set of IV points for a new Pokemon.
-     * 
-     * @returns A random set of IV points.
-     */
-    newPokemonIVs: (this: imathdecidr.IMathDecidr) => { [i: string]: number };
-
-    /**
-     * Computes a blank set of EV points for a new Pokemon.
-     * 
-     * @returns A blank set of EV points.
-     */
-    newPokemonEVs: (this: imathdecidr.IMathDecidr) => { [i: string]: number };
-
-    /**
-     * Computes how much experience a new Pokemon should start with, based on its
-     * type and level.
-     * 
-     * @param title   The type of Pokemon.
-     * @param level   The level of the Pokemon.
-     * @returns How much experience the Pokemon should start with.
-     */
-    newPokemonExperience: (this: imathdecidr.IMathDecidr, title: string[], level: number) => ibattlemovr.IActorExperience;
-
-    /**
-     * Computes a Pokemon's new statistic based on its IVs and EVs.
-     * 
-     * @param statistic   Which statistic to compute.
-     * @returns A new value for the statistic.
-     */
-    pokemonStatistic: (this: imathdecidr.IMathDecidr, pokemon: IPokemon, statistic: string) => number;
-
-    /**
-     * Determines whether a wild encounter should occur when walking through grass.
-     * 
-     * @param grass   The grass Thing being walked through.
-     * @returns Whether a wild encounter should occur.
-     */
-    doesGrassEncounterHappen: (this: imathdecidr.IMathDecidr, grass: IGrass) => boolean;
-
-    /**
-     * Determines whether a Pokemon may be caught by a ball.
-     * 
-     * @param pokemon   The Pokemon the ball is attempting to catch.
-     * @param ball   The ball attempting to catch the Pokemon.
-     * @returns Whether the Pokemon may be caught.
-     */
-    canCatchPokemon: (this: imathdecidr.IMathDecidr, pokemon: IPokemon, ball: IBattleBall) => boolean;
-
-    /**
-     * Determines whether the player may flee a wild Pokemon encounter.
-     * 
-     * @param pokemon   The player's current Pokemon.
-     * @param enemy   The wild Pokemon.
-     * @param battleInfo   Information on the current battle.
-     * @returns Whether the player may flee.
-     */
-    canEscapePokemon: (this: imathdecidr.IMathDecidr, pokemon: IPokemon, enemy: IPokemon, battleInfo: IBattleInfo) => boolean;
-
-    /**
-     * Calculates how many times a failed Pokeball should shake.
-     * 
-     * @param pokemon   The wild Pokemon the ball is failing to catch.
-     * @param ball   The Pokeball attempting to catch the wild Pokemon.
-     * @returns How many times the balls hould shake.
-     */
-    numBallShakes: (this: imathdecidr.IMathDecidr, pokemon: IPokemon, ball: IBattleBall) => number;
-
-    /**
-     * Determines what move an opponent should take in battle.
-     * 
-     * @param player   The in-battle player.
-     * @param opponent   The in-battle opponent.
-     * @returns The contatenated name of the move the opponent will choose.
-     */
-    opponentMove: (
-        this: imathdecidr.IMathDecidr,
-        player: IBattler,
-        opponent: IBattler) => string;
-
-    /**
-     * Checks whether a Pokemon contains any of the given types.
-     * 
-     * @param pokemon   A Pokemon.
-     * @param types   The types to check.
-     * @returns Whether the Pokemon's types includes any of the given types.
-     */
-    pokemonMatchesTypes: (this: imathdecidr.IMathDecidr, pokemon: IPokemon, types: string[]) => boolean;
-
-    /**
-     * Checks whether a move only has a status effect (does no damage, or nothing).
-     * 
-     * @param move   The move.
-     * @returns Whether the moves has only a status effect.
-     */
-    moveOnlyStatuses: (this: imathdecidr.IMathDecidr, move: IMoveSchema) => boolean;
-
-    /**
-     * Modifies a move possibility's priority based on battle state.
-     * 
-     * @param possibility   A move possibility.
-     * @param modification   A modification summary for a part of the battle state.
-     * @param target   The Pokemon being targeted.
-     * @param amount   How much to modify the move's priority.
-     */
-    applyMoveEffectPriority: (possibility: IMovePossibility, modification: IBattleModification, target: IPokemon, amount: number) => void;
-
-    /**
-     * Determines whether a player's Pokemon should move before the opponent's.
-     * 
-     * @param player   The in-battle player.
-     * @param choicePlayer   The concatenated name of the move the player chose.
-     * @param opponent   The in-battle opponent.
-     * @param choiesOpponent   The concatenated name of the move the opponent chose.
-     * @returns Whether the player will move before the opponent.
-     */
-    playerMovesFirst: (player: IBattler, choicePlayer: string, opponent: IBattler, choiceOpponent: string) => boolean;
-
-    /**
-     * Computes how much damage a move should do to a Pokemon.
-     * 
-     * @param move   The concatenated name of the move.
-     * @param attacker   The attacking pokemon.
-     * @param defender   The defending Pokemon.
-     * @returns How much damage should be dealt.
-     */
-    damage: (this: imathdecidr.IMathDecidr, move: string, attacker: IPokemon, defender: IPokemon) => number;
-
-    /**
-     * Determines the damage modifier against a defending Pokemon.
-     * 
-     * @param move   The attacking Pokemon's move.
-     * @param attacker   The attacking Pokemon.
-     * @param defender   The defending Pokemon.
-     * @returns The damage modifier, as a multiplication constant.
-     */
-    damageModifier: (move: IMoveSchema, attacker: IPokemon, defender: IPokemon) => number;
-
-    /**
-     * Determines whether a move should be a critical hit.
-     * 
-     * @param move   The concatenated name of the move.
-     * @param attacker   The attacking Pokemon.
-     * @returns Whether the move should be a critical hit.
-     */
-    criticalHit: (this: imathdecidr.IMathDecidr, move: string, attacker: IPokemon) => boolean;
-
-    /**
-     * Determines the type effectiveness of a move on a defending Pokemon.
-     * 
-     * @param move   The concatenated name of the move.
-     * @param defender   The defending Pokemon.
-     * @returns A damage modifier, as a multiplication constant.
-     */
-    typeEffectiveness: (this: imathdecidr.IMathDecidr, move: string, defender: IPokemon) => number;
-
-    /**
-     * Computes how much experience a new Pokemon should start with.
-     * 
-     * @param title   The name of the Pokemon.
-     * @param level   The level of the Pokemon.
-     * @returns An amount of experience.
-     */
-    experienceStarting: (this: imathdecidr.IMathDecidr, title: string[], level: number) => number;
-
-    /**
-     * Computes how much experience should be gained from defeating a Pokemon.
-     * 
-     * @returns How much experience is to be gained.
-     * @remarks This will need to be changed to accomodate rewarding multiple Pokemon.
-     */
-    experienceGained: (this: imathdecidr.IMathDecidr, player: IBattler, opponent: IBattler) => number;
-
-    /**
-     * Computes how wide a health bar should be.
-     * 
-     * @param widthFullBar   The maximum possible width.
-     * @param hp   How much HP a Pokemon currently has.
-     * @param hpNormal   The maximum HP the Pokemon may have.
-     * @returns How wide the health bar should be.
-     */
-    widthHealthBar: (widthFullBar: number, hp: number, hpNormal: number) => number;
-}
-
-/**
  * FullScreenPokemon-specific IMathDecidr usage.
  */
 export interface IFullScreenPokemonMathDecidr extends imathdecidr.IMathDecidr {
@@ -457,7 +234,7 @@ export interface IFullScreenPokemonMathDecidr extends imathdecidr.IMathDecidr {
     /**
      * Stored equations bound to the IMathDecidr.
      */
-    readonly equations: IMathEquations;
+    readonly equations: {};
 }
 
 /**
