+++ src/Settings/Math.ts
@@ -1,15 +1,7 @@
-import { IActorExperience, IMove } from "battlemovr/lib/IBattleMovr";
-
 import { Actions } from "../components/Actions";
 import { Cycling } from "../components/Cycling";
 import { Fishing } from "../components/Fishing";
-import {
-    IBattleBall, IBattleInfo, IBattleModification, IBattler,
-    ICharacter, IFullScreenPokemonMathDecidr, IGrass, IHMMoveSchema,
-    IMathModuleSettings, IMovePossibility, IMoveSchema, IPokemon,
-    IPokemonListing, IPokemonMoveListing, IRod,
-    IWildPokemonSchema
-} from "../IFullScreenPokemon";
+import { IHMMoveSchema, IMathModuleSettings, IRod, } from "../IFullScreenPokemon";
 
 /* tslint:disable:max-line-length */
 /* tslint:disable object-literal-key-quotes */
@@ -18,527 +10,7 @@ export function GenerateMathSettings(): IMathModuleSettings {
     "use strict";
 
     return {
-        "equations": {
-            "averageLevel": function (this: IFullScreenPokemonMathDecidr, actors: IPokemon[]): number {
-                let average: number = 0;
-
-                for (let i: number = 0; i < actors.length; i += 1) {
-                    average += actors[i].level;
-                }
-
-                return Math.round(average / actors.length);
-            },
-            "averageLevelWildPokemon": function (this: IFullScreenPokemonMathDecidr, options: IWildPokemonSchema[]): number {
-                let average: number = 0;
-
-                for (const wildPokemon of options) {
-                    if (wildPokemon.level) {
-                        average += wildPokemon.level * wildPokemon.rate;
-                        continue;
-                    }
-
-                    if (!wildPokemon.levels) {
-                        throw new Error("Wild Pokemon must have wither .level of .levels defined.");
-                    }
-
-                    let levelAverage: number = 0;
-
-                    for (const level of wildPokemon.levels) {
-                        levelAverage += level * (1 / wildPokemon.levels.length);
-                    }
-
-                    average += levelAverage * wildPokemon.rate;
-                }
-
-                return Math.round(average);
-            },
-            "speedCycling": function (this: IFullScreenPokemonMathDecidr, thing: ICharacter): number {
-                return thing.speed * 2;
-            },
-            "speedWalking": function (this: IFullScreenPokemonMathDecidr, thing: ICharacter): number {
-                return Math.round(8 / thing.speed);
-            },
-            "newPokemon": function (this: IFullScreenPokemonMathDecidr, title: string[], level?: number, moves?: IMove[], iv?: number, ev?: number): IPokemon {
-                const statisticNames: string[] = this.constants.statisticNames;
-                const pokemon: any = {
-                    "title": title,
-                    "nickname": title,
-                    "level": level || 1,
-                    "moves": moves || this.compute("newPokemonMoves", title, level || 1),
-                    "types": this.constants.pokemon[title.join("")].types,
-                    "status": "",
-                    "IV": iv || this.compute("newPokemonIVs"),
-                    "EV": ev || this.compute("newPokemonEVs"),
-                    "experience": this.compute("newPokemonExperience", title, level || 1)
-                };
-
-                for (const statistic of statisticNames) {
-                    pokemon[statistic] = this.compute("pokemonStatistic", pokemon, statistic);
-                    pokemon[statistic + "Normal"] = pokemon[statistic];
-                }
-
-                return pokemon;
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/XXXXXXX_(Pok%C3%A9mon)/Generation_I_learnset
-            "newPokemonMoves": function (this: IFullScreenPokemonMathDecidr, title: string[], level: number): IMove[] {
-                const possibilities: IPokemonMoveListing[] = this.constants.pokemon[title.join("")].moves.natural;
-                const output: IMove[] = [];
-                let move: IPokemonMoveListing;
-                let newMove: IMove;
-                let end: number;
-
-                for (end = 0; end < possibilities.length; end += 1) {
-                    if (possibilities[end].level > level) {
-                        break;
-                    }
-                }
-
-                for (let i: number = Math.max(end - 4, 0); i < end; i += 1) {
-                    move = possibilities[i];
-                    newMove = {
-                        title: move.move,
-                        remaining: this.constants.moves[move.move].PP,
-                        uses: this.constants.moves[move.move].PP
-                    };
-
-                    output.push(newMove);
-                }
-
-                return output;
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Individual_values
-            "newPokemonIVs": function (this: IFullScreenPokemonMathDecidr): { [i: string]: number } {
-                const attack: number = this.constants.numberMaker.randomIntWithin(0, 15);
-                const defense: number = this.constants.numberMaker.randomIntWithin(0, 15);
-                const speed: number = this.constants.numberMaker.randomIntWithin(0, 15);
-                const special: number = this.constants.numberMaker.randomIntWithin(0, 15);
-                const output: any = {
-                    "Attack": attack,
-                    "Defense": defense,
-                    "Speed": speed,
-                    "Special": special
-                };
-
-                output.HP = (
-                    8 * (attack % 2)
-                    + 4 * (defense % 2)
-                    + 2 * (speed % 2)
-                    + (special % 2)
-                );
-
-                return output;
-            },
-            "newPokemonEVs": function (this: IFullScreenPokemonMathDecidr): { [i: string]: number } {
-                return {
-                    "Attack": 0,
-                    "Defense": 0,
-                    "Speed": 0,
-                    "Special": 0
-                };
-            },
-            "newPokemonExperience": function (this: IFullScreenPokemonMathDecidr, title: string[], level: number): IActorExperience {
-                return {
-                    current: this.compute("experienceStarting", title, level),
-                    next: this.compute("experienceStarting", title, level + 1)
-                };
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Individual_values
-            // Note: the page mentions rounding errors... 
-            "pokemonStatistic": function (this: IFullScreenPokemonMathDecidr, pokemon: IPokemon, statistic: string): number {
-                let topExtra: number = 0;
-                let added: number = 5;
-                let base: number = (this.constants.pokemon[pokemon.title.join("")] as any)[statistic];
-                let iv: number = (pokemon.IV as any)[statistic] || 0;
-                let ev: number = (pokemon.EV as any)[statistic] || 0;
-                let level: number = pokemon.level;
-
-                if (statistic === "HP") {
-                    topExtra = 50;
-                    added = 10;
-                }
-
-                let numerator: number = (iv + base + (Math.sqrt(ev) / 8) + topExtra) * level;
-
-                return (numerator / 50 + added) | 0;
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Tall_grass
-            "doesGrassEncounterHappen": function (this: IFullScreenPokemonMathDecidr, grass: IGrass): boolean {
-                return this.constants.numberMaker.randomBooleanFraction(grass.rarity, 187.5);
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Catch_rate#Capture_method_.28Generation_I.29
-            "canCatchPokemon": function (this: IFullScreenPokemonMathDecidr, pokemon: IPokemon, ball: IBattleBall): boolean {
-                // 1. If a Master Ball is used, the Pokemon is caught.
-                if (ball.type === "Master") {
-                    return true;
-                }
-
-                // 2. Generate a random number, N, depending on the type of ball used.
-                let n: number = this.constants.numberMaker.randomInt(ball.probabilityMax);
-
-                // 3. The Pokemon is caught if...
-                if (pokemon.status) { // ... it is asleep or frozen and N is less than 25.
-                    if (n < 25) {
-                        if (this.constants.statuses.probability25[pokemon.status]) {
-                            return true;
-                        }
-                    } else if (n < 12) { // ... it is paralyzed, burned, or poisoned and N is less than 12.
-                        if (this.constants.statuses.probability12[pokemon.status]) {
-                            return true;
-                        }
-                    }
-                }
-
-                // 4. Otherwise, if N minus the status value is greater than the Pokemon's catch rate, the Pokemon breaks free.
-                if (n - this.constants.statuses.levels[pokemon.status] > pokemon.catchRate) {
-                    return false;
-                }
-
-                // 5. If not, generate a random value, M, between 0 and 255.
-                let m: number = this.constants.numberMaker.randomInt(255);
-
-                // 6. Calculate f.
-                let f: number = Math.max(
-                    Math.min(
-                        (pokemon.HPNormal * 255 * 4) | 0 / (pokemon.HP * ball.rate) | 0,
-                        255),
-                    1);
-
-                // 7. If f is greater than or equal to M, the Pokemon is caught. Otherwise, the Pokemon breaks free.
-                return f > m;
-            },
-            // @todo Add functionality.
-            "canLandFish": function (this: IFullScreenPokemonMathDecidr): boolean {
-                return true;
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Escape#Generation_I_and_II
-            "canEscapePokemon": function (this: IFullScreenPokemonMathDecidr, pokemon: IPokemon, enemy: IPokemon, battleInfo: IBattleInfo): boolean {
-                const a: number = pokemon.Speed;
-                const b: number = (enemy.Speed / 4) % 256;
-                const c: number = battleInfo.currentEscapeAttempts || 0;
-                const f: number = (a * 32) / b + 30 * c;
-
-                if (f > 255 || b === 0) {
-                    return true;
-                }
-
-                return this.constants.numberMaker.randomInt(256) < f;
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Catch_rate#Capture_method_.28Generation_I.29
-            "numBallShakes": function (this: IFullScreenPokemonMathDecidr, pokemon: IPokemon, ball: IBattleBall): number {
-                // 1. Calculate d.
-                const d: number = pokemon.catchRate * 100 / ball.rate;
-
-                // 2. If d is greater than or equal to 256, the ball shakes three times before the Pokemon breaks free.
-                if (d >= 256) {
-                    return 3;
-                }
-
-                // 3. If not, calculate x = d * f / 255 + s, where s is 10 if the Pokemon is asleep or frozen or 5 if it is paralyzed, poisoned, or burned.
-                const f: number = Math.max(
-                    Math.min(
-                        (pokemon.HPNormal * 255 * 4) | 0 / (pokemon.HP * ball.rate) | 0,
-                        255),
-                    1);
-                const x: number = d * f / 255 + this.constants.statuses.shaking[pokemon.status];
-
-                // 4. If... 
-                if (x < 10) { // x < 10: the Ball misses the Pokemon completely.
-                    return 0;
-                }
-
-                // x < 30: the Ball shakes once before the Pokemon breaks free.
-                if (x < 30) {
-                    return 1;
-                }
-
-                // x < 70: the Ball shakes twice before the Pokemon breaks free.
-                if (x < 70) {
-                    return 2;
-                }
-
-                // Otherwise, the Ball shakes three times before the Pokemon breaks free.
-                return 3;
-            },
-            // http://wiki.pokemonspeedruns.com/index.php/Pok%C3%A9mon_Red/Blue/Yellow_Trainer_AI
-            // TO DO: Also filter for moves with > 0 remaining remaining...
-            "opponentMove": function (this: IFullScreenPokemonMathDecidr, player: IBattler, opponent: IBattler): string {
-                let possibilities: IMovePossibility[] = opponent.selectedActor!.moves.map(
-                    (move: IMove): IMovePossibility => {
-                        return {
-                            "move": move.title,
-                            "priority": 10
-                        };
-                    });
-
-                // Wild Pokemon just choose randomly
-                if (opponent.category === "Wild") {
-                    return this.constants.numberMaker.randomArrayMember(possibilities).move;
-                }
-
-                // Modification 1: Do not use a move that only statuses (e.g. Thunder Wave) if the player's pokémon already has a status.
-                if (player.selectedActor!.status && !opponent.dumb) {
-                    for (const possibility of possibilities) {
-                        if (this.compute("moveOnlyStatuses", possibility.move)) {
-                            possibility.priority += 5;
-                        }
-                    }
-                }
-
-                // Modification 2: On the second turn the pokémon is out, prefer a move with one of the following effects...
-                if (this.compute("pokemonMatchesTypes", opponent, this.constants.battleModifications["Turn 2"])) {
-                    for (const possibility of possibilities) {
-                        this.compute(
-                            "applyMoveEffectPriority",
-                            possibility,
-                            this.constants.battleModifications["Turn 2"],
-                            player.selectedActor,
-                            1);
-                    }
-                }
-
-                // Modification 3 (Good AI): Prefer a move that is super effective. Do not use moves that are not very effective as long as there is an alternative.
-                if (this.compute("pokemonMatchesTypes", opponent, this.constants.battleModifications["Good AI"])) {
-                    for (let i: number = 0; i < possibilities.length; i += 1) {
-                        this.compute(
-                            "applyMoveEffectPriority",
-                            possibilities[i],
-                            this.constants.battleModifications["Good AI"],
-                            player.selectedActor,
-                            1);
-                    }
-                }
-
-                // The AI uses rejection sampling on the four moves with ratio 63:64:63:66, with only the moves that are most favored after applying the modifications being acceptable.
-                let lowest: number = possibilities[0].priority;
-                if (possibilities.length > 1) {
-                    for (let i: number = 1; i < possibilities.length; i += 1) {
-                        if (possibilities[i].priority < lowest) {
-                            lowest = possibilities[i].priority;
-                        }
-                    }
-                    possibilities = possibilities.filter(function (possibility: IMovePossibility): boolean {
-                        return possibility.priority === lowest;
-                    });
-                }
-
-                return this.constants.numberMaker.randomArrayMember(possibilities).move;
-            },
-            "pokemonMatchesTypes": function (this: IFullScreenPokemonMathDecidr, pokemon: IPokemon, types: string[]): boolean {
-                for (let i: number = 0; i < types.length; i += 1) {
-                    if (pokemon.types.indexOf(types[i]) !== -1) {
-                        return true;
-                    }
-                }
-
-                return false;
-            },
-            "moveOnlyStatuses": function (this: IFullScreenPokemonMathDecidr, move: IMoveSchema): boolean {
-                return move.damage === "Non-Damaging" && move.effect === "Status";
-            },
-            "applyMoveEffectPriority": function (this: IFullScreenPokemonMathDecidr, possibility: IMovePossibility, modification: IBattleModification, target: IPokemon, amount: number): void {
-                const preferences: ([string, string, number] | [string, string])[] = modification.preferences;
-                const move: IMoveSchema = this.constants.moves[possibility.move];
-
-                for (let i: number = 0; i < preferences.length; i += 1) {
-                    let preference: [string, string, number] | [string, string] = preferences[i];
-
-                    switch (preference[0]) {
-                        // ["Move", String]
-                        // Favorable match
-                        case "Move":
-                            if (possibility.move === preference[1]) {
-                                possibility.priority -= amount;
-                                return;
-                            }
-                            break;
-
-                        // ["Raise", String, Number]
-                        // Favorable match
-                        case "Raise":
-                            if (
-                                move.effect === "Raise"
-                                && move.raise === preference[1]
-                                && move.amount === preference[2]
-                            ) {
-                                possibility.priority -= amount;
-                                return;
-                            }
-                            break;
-
-                        // ["Lower", String, Number]
-                        // Favorable match
-                        case "Lower":
-                            if (
-                                move.effect === "Lower"
-                                && move.lower === preference[1]
-                                && move.amount === preference[2]
-                            ) {
-                                possibility.priority -= amount;
-                                return;
-                            }
-                            break;
-
-                        // ["Super", String, String]
-                        // Favorable match
-                        case "Super":
-                            if (
-                                move.damage !== "Non-Damaging"
-                                && move.type === preference[0]
-                                && target.types.indexOf(preference[1]) !== -1
-                            ) {
-                                possibility.priority -= amount;
-                                return;
-                            }
-                            break;
-
-                        // ["Weak", String, String]
-                        // Unfavorable match
-                        case "Weak":
-                            if (
-                                move.damage !== "Non-Damaging"
-                                && move.type === preference[0]
-                                && target.types.indexOf(preference[1]) !== -1
-                            ) {
-                                possibility.priority += amount;
-                                return;
-                            }
-                            break;
-
-                        // By default, do nothing
-                        default:
-                            break;
-                    }
-                }
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Priority
-            // TO DO: Account for items, switching, etc.
-            // TO DO: Factor in spec differences from paralyze, etc.
-            "playerMovesFirst": function (this: IFullScreenPokemonMathDecidr, player: IBattler, choicePlayer: string, opponent: IBattler, choiceOpponent: string): boolean {
-                const movePlayer: IMoveSchema = this.constants.moves[choicePlayer];
-                const moveOpponent: IMoveSchema = this.constants.moves[choiceOpponent];
-
-                if (movePlayer.priority === moveOpponent.priority) {
-                    return player.selectedActor!.Speed > opponent.selectedActor!.Speed;
-                }
-
-                return movePlayer.priority > moveOpponent.priority;
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Damage#Damage_formula
-            // http://bulbapedia.bulbagarden.net/wiki/Critical_hit
-            // TO DO: Factor in spec differences from burns, etc.
-            "damage": function (this: IFullScreenPokemonMathDecidr, move: string, attacker: IPokemon, defender: IPokemon): number {
-                let base: string | number = this.constants.moves[move].power;
-
-                // A base attack that's not numeric means no damage, no matter what
-                if (!base || isNaN(base as number)) {
-                    return 0;
-                }
-
-                // Don't bother calculating infinite damage: it's going to be infinite
-                if (base === Infinity) {
-                    return Infinity;
-                }
-
-                const critical: boolean = this.compute("criticalHit", move, attacker);
-                const level: number = attacker.level * Number(critical);
-                const attack: number = attacker.Attack;
-                const defense: number = defender.Defense;
-                const modifier: number = this.compute("damageModifier", move, attacker, defender);
-
-                return Math.round(
-                    Math.max(
-                        ((((2 * level + 10) / 250) * (attack / defense) * (base as number) + 2) | 0) * modifier,
-                        1));
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Damage#Damage_formula
-            // http://bulbapedia.bulbagarden.net/wiki/Critical_hit
-            "damageModifier": function (this: IFullScreenPokemonMathDecidr, move: IMoveSchema, attacker: IPokemon, defender: IPokemon): number {
-                const stab: number = attacker.types.indexOf(move.type) !== -1 ? 1.5 : 1;
-                const type: number = this.compute("typeEffectiveness", move, defender);
-
-                return stab * type * this.constants.numberMaker.randomWithin(.85, 1);
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Critical_hit
-            "criticalHit": function (this: IFullScreenPokemonMathDecidr, move: string, attacker: IPokemon): boolean {
-                const moveInfo: IMoveSchema = this.constants.moves[move];
-                const baseSpeed: number = this.constants.pokemon[attacker.title.join("")].Speed;
-                let denominator: number = 512;
-
-                // Moves with a high critical-hit ratio, such as Slash, are eight times more likely to land a critical hit, resulting in a probability of BaseSpeed / 64.
-                if (moveInfo.criticalRaised) {
-                    denominator /= 8;
-                }
-
-                // "Focus Energy and Dire Hit were intended to increase the critical hit rate, ..."
-                // In FullScreenPokemon, they work as intended! Fans who prefer the
-                // original behavior are free to fork the repo. As the original
-                // behavior is a glitch (and conflicts with creators' intentions),
-                // it is not duplicated here.
-                if (attacker.criticalHitProbability) {
-                    denominator /= 4;
-                }
-
-                // As with move accuracy in the handheld games, if the probability of landing a critical hit would be 100%, it instead becomes 255/256 or about 99.6%.
-                return this.constants.numberMaker.randomBooleanProbability(Math.max(baseSpeed / denominator, 255 / 256));
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Type/Type_chart#Generation_I
-            "typeEffectiveness": function (this: IFullScreenPokemonMathDecidr, move: string, defender: IPokemon): number {
-                const defenderTypes: string[] = this.constants.pokemon[defender.title.join("")].types;
-                const moveIndex: number = this.constants.types.indices[this.constants.moves[move].type];
-                let total: number = 1;
-
-                for (let i: number = 0; i < defenderTypes.length; i += 1) {
-                    total *= this.constants.types.table[moveIndex][this.constants.types.indices[defenderTypes[i]]];
-                }
-
-                return total;
-            },
-            // http://m.bulbapedia.bulbagarden.net/wiki/Experience#Relation_to_level
-            // Wild Pokémon of any level will always have the base amount of experience required to reach that level when caught, as will Pokémon hatched from Eggs.
-            "experienceStarting": function (this: IFullScreenPokemonMathDecidr, title: string[], level: number): number {
-                let reference: IPokemonListing = this.constants.pokemon[title.join("")];
-
-                // TODO: remove defaulting to mediumFast
-                switch (reference.experienceType) {
-                    case "fast":
-                        return (4 * Math.pow(level, 3)) / 5;
-                    case "mediumSlow":
-                        return (
-                            (6 / 5) * Math.pow(level, 3)
-                            - (15 * Math.pow(level, 2))
-                            + (100 * level)
-                            - 140);
-                    case "slow":
-                        return (5 * Math.pow(level, 3)) / 4;
-                    // case "mediumFast":
-                    default:
-                        return Math.pow(level, 3);
-                }
-            },
-            // http://bulbapedia.bulbagarden.net/wiki/Experience#Gain_formula
-            "experienceGained": function (this: IFullScreenPokemonMathDecidr, player: IBattler, opponent: IBattler): number {
-                // a is equal to 1 if the fainted Pokemon is wild, or 1.5 if the fainted Pokemon is owned by a Trainer
-                let a: number = opponent.category === "Trainer" ? 1.5 : 1;
-
-                // b is the base experience yield of the fainted Pokemon's species
-                let b: number = 64; // (Bulbasaur) TO DO: add this in
-
-                // lf is the level of the fainted Pokemon
-                let lf: number = opponent.selectedActor!.level;
-
-                // s is equal to (in Gen I), if Exp. All is not in the player's Bag...
-                // TO DO: Account for modifies like Exp. All
-                let s: number = 1;
-
-                // t is equal to 1 if the winning Pokemon's curent owner is its OT, or 1.5 if the Pokemon was gained in a domestic trade
-                let t: number = player.selectedActor!.traded ? 1.5 : 1;
-
-                return (((a * t * b * lf) | 0) / ((7 * s) | 0)) | 0;
-            },
-            "widthHealthBar": function (this: IFullScreenPokemonMathDecidr, widthFullBar: number, hp: number, hpNormal: number): number {
-                return (widthFullBar - 1) * hp / hpNormal;
-            }
-        },
+        "equations": {},
         "constants": {
             "statisticNames": ["HP", "Attack", "Defense", "Speed", "Special"],
             "statisticNamesDisplayed": ["Attack", "Defense", "Speed", "Special"],
